watchers/requirements.txt
# Python dependencies for AI Employee Watchers
# Install with: pip install -r requirements.txt

# File system monitoring
watchdog>=4.0.0

# Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Browser automation (LinkedIn, Facebook)
playwright>=1.40.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (for Facebook Graph API, etc.)
requests>=2.31.0

# MCP SDK
mcp>=1.0.0
